requirements.txt
web3==6.10.0
eth-account==0.10.0
eth-keys==0.5.0
requests==2.31.0
numpy==1.24.3
rich==13.5.2
python-dotenv==1.0.0
flask==3.0.0
flask-cors==4.0.0
anthropic>=0.25.0
websockets>=12.0
groq>=0.4.0
praw>=7.7.0
streamlit>=1.28.0
plotly>=5.17.0
pandas>=2.0.0
